01-agent-pipeline/pyproject.toml
[project]
name = "beacon-pipeline"
version = "0.1.0"
description = "Beacon Agent Pipeline - AI research agent that builds knowledge bases"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "tavily-python>=0.5.0",
    "trafilatura>=2.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["beacon"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
